refactor: Revamp settings page structure and components
- Reorganized the settings page by removing unused state and logic related to password change, alternative email, notification, and privacy settings.
- Introduced new components for ChangePassword, AlternativeEmail, NotificationSettings, PrivacySettings, AccountSettings, SupportHelp, and AppearanceSettings to enhance modularity and maintainability.
- Updated the rendering logic to utilize the new components, improving code clarity and separation of concerns.
